+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Improved error messages
 * Defended `cluster.balancer` with `sync.RWMutex` on `cluster.Insert`, `cluster.Update`, `cluster.Remove` and `cluster.Get`
 * Excluded `Close` and `Park` methods from `conn.Conn` interface
+* Fixed bug with `Create()` on multi-balancer 
 
 ## 3.13.0
 * Refactored `Connection` interface